pyproject.toml
[project]
name = "notion-game-list"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.3",
    "dateparser>=1.2.2",
    "pyinstaller>=6.16.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "termcolor>=1.1.0",
]
